fix: loading page, Slpash screen created
