+++ go-for-eat/screens/Splash/styles.js
@@ -1,0 +1,15 @@
+import { StyleSheet } from 'react-native';
+
+export default StyleSheet.create({
+  container:{
+    flex: 1,
+    backgroundColor: '#2ECC71',
+    justifyContent: 'center',
+    alignItems: 'center'
+  },
+  image: {
+    transform: [
+      {'scale': 0.6}
+    ]
+  }
+});
